+++ src/main/java/com/schwab/urlshortener/service/UrlShorteningService.java
@@ -1,0 +1,9 @@
+package com.schwab.urlshortener.service;
+
+import com.schwab.urlshortener.dto.CreateShortUrlRequest;
+import com.schwab.urlshortener.dto.ShortUrlResponse;
+
+public interface UrlShorteningService {
+
+    ShortUrlResponse createShortUrl(CreateShortUrlRequest request);
+}